LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/multicol-fill-balance-021-expected.txt


FAIL Ruby text should affect column height assert_equals: expected 100 but got 70
